refactor: split engine.go, deduplicate logic, add SDKScanner and tests
Split engine.go (2045→878 lines) into 7 focused files:
- webprobe.go: web service probing
- pedm.go: performance monitoring (PEDM)
- stage_fingerprint.go: fingerprint execution
- filter.go: PoC skipping/filtering logic
- oob_coordinator.go: OOB adapter + manager initialization
- context.go: ScanContext for runtime callback separation
- errors.go: structured error types for scan operations

Deduplications:
- ContainsOOBToken/PocUsesOOB exported via config package
- IsNetOnly() method on poc.Poc (eliminates duplicate closures)
- TargetStrings() on Options (4 duplicate loops eliminated)
- ApplyDefaults()/ValidateRateLimitModes() helpers
- sendOrDrop[T] generic channel helper (6 call sites)

SDK:
- RunnerOption pattern with WithSDKMode()
- afrog.go: SDKScanner public API with Run/Stop/GetResults
- shutdown.go, progress_render.go: pending integration types

Tests: 43 new cases across filter, context, poc, oob_coordinator, fingerprint

Verification:
  go build ./...  ✅
  go vet ./...    ✅
  go test ./...   ✅ (12 packages)
  gofmt -l .      ✅ (0 violations)
