+++ pkg/config/options_legacy_oob_test.go
@@ -76,4 +76,3 @@ expression: r0()
 		})
 	}
 }
-
